backend/requirements.txt
# 🦅 Harpia - Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# OpenAI
openai>=1.50.0
openai-agents>=0.1.0
chatkit>=0.1.0

# Supabase
supabase>=2.0.0

# HTTP Client
httpx>=0.27.0

# Google Gemini
google-generativeai>=0.8.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
